requirements.txt
selenium>=4.15.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
PyYAML>=6.0.1
